+++ src/shared/components/palette/organisation/write-organisation-field.component.ts
@@ -90,8 +90,26 @@ export class WriteOrganisationFieldComponent extends AbstractFieldWriteComponent
     if (organisation.postCode && organisation.postCode.toLowerCase().includes(this.trimAll(lowerOrgSearchText))) {
       return true;
     }
-    // noinspection RedundantIfStatementJS
     if (organisation.name && organisation.name.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    if (organisation.addressLine1 && organisation.addressLine1.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    if (organisation.addressLine2 && organisation.addressLine2.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    if (organisation.addressLine3 && organisation.addressLine3.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    if (organisation.townCity && organisation.townCity.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    if (organisation.county && organisation.county.toLowerCase().includes(lowerOrgSearchText)) {
+      return true;
+    }
+    // noinspection RedundantIfStatementJS
+    if (organisation.country && organisation.country.toLowerCase().includes(lowerOrgSearchText)) {
       return true;
     }
     return false;