+++ RELEASE-NOTES.md
@@ -1,4 +1,7 @@
 ## RELEASE NOTES
+### Version 2.64.37-organisation-complex-field-type
+**EUI-2087** - Search organisation support full UK address fields
+
 ### Version 2.64.36-hide-generic-callback-error-message
 **EUI-2344** - Hide generic callback error message when there is a custom configured message provided by a Service